move templates into highlight/view dir
