+++ src/Highlight/View/index.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "../Highlight/Components/App";
+import Extension from "./Extension";
 import './style.css'
 
 const root = ReactDOM.createRoot(document.getElementById("root")!);
 
-root.render(<App/>);
+root.render(<Extension/>);
